# bookworm-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookworm API",
        "description": "Bookworm API documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/best-sellers": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get best selling books",
                "description": "",
                "operationId": "ce4b208024cd0b674ef43cac5bcd3f98",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term to filter the results by",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Number of items to return per page. Defaults to 20",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "cache",
                        "in": "query",
                        "description": "Whether to retrieve data from cache or API. Defaults to true",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "List all books",
                "operationId": "3648e62dc9941ff0dd4d9f6c762037de",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search string",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Page size",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "operationId": "57821293831aea7c943b685c527e097a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a book",
                "operationId": "6eaf879d509af85a0ba27d3d57abde81",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update a book",
                "operationId": "9a19187021c7baf6360321f64024b865",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete a book",
                "operationId": "dda2b30434da8328795d718061be7795",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            }
        },
        "/api/books-favourites": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get user's favourite books",
                "description": "",
                "operationId": "57503a0e35fbbc17cbe8771dec97a881",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books/{bookId}/favourite": {
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Toggle book favourite status",
                "description": "",
                "operationId": "d0f0a8740e276bbc2b126b4ba185c216",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "type": {
                                            "type": "string",
                                            "enum": [
                                                "added",
                                                "removed"
                                            ],
                                            "example": "added"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Book": {
                "title": "Book",
                "description": "Book model",
                "properties": {
                    "externalId": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "authors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "description": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "link": {
                        "type": "string"
                    },
                    "rating": {
                        "type": "integer"
                    },
                    "price": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "BestSellersRequest": {
                "title": "BestSellersRequest",
                "description": "Request to get best selling books",
                "required": [
                    "search",
                    "pageSize"
                ],
                "properties": {
                    "search": {
                        "type": "string",
                        "example": "Harry Potter"
                    },
                    "pageSize": {
                        "type": "integer",
                        "example": 20
                    }
                },
                "type": "object"
            },
            "BookCreateRequest": {
                "required": [
                    "externalId",
                    "title",
                    "authors",
                    "description",
                    "image"
                ],
                "properties": {
                    "externalId": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "authors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "description": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string",
                        "format": "url"
                    },
                    "link": {
                        "type": "string",
                        "format": "url"
                    }
                },
                "type": "object"
            },
            "BookUpdateRequest": {
                "required": [
                    "price",
                    "rating"
                ],
                "properties": {
                    "price": {
                        "type": "number",
                        "format": "float"
                    },
                    "rating": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Books",
            "description": "Books"
        }
    ]
}